functions/requirements.txt
azure-functions
azure-storage-queue
azure-core
python-dotenv
pydantic
requests
typing-extensions
isoscore